Moves collection list component
